Extra Merge Doubles in Edit Mode tool
